+++ NOICommunityLib/Sources/VimeoOEmbedClient/VimeoOEmbedError.swift
@@ -1,0 +1,18 @@
+// SPDX-FileCopyrightText: NOI Techpark <digital@noi.bz.it>
+//
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+//
+//  VimeoOEmbedError.swift
+//  VimeoOEmbedClient
+//
+//  Created by Matteo Matassoni on 13/01/25.
+//
+
+import Foundation
+
+// MARK: - VimeoOEmbedError
+
+public enum VimeoOEmbedError: Error {
+	case invalidURL
+}
