+++ NOICommunityLib/Sources/VimeoVideoThumbnailClient/Interfaces/VimeoVideoThumbnailClient.swift
@@ -4,13 +4,15 @@
 
 //
 //  VimeoVideoThumbnailClient.swift
-//  
+//  VimeoVideoThumbnailClient
 //
 //  Created by Camilla on 23/12/24.
 //
 
 
 import Foundation
+
+// MARK: - VimeoVideoThumbnailClient
 
 public protocol VimeoVideoThumbnailClient {
